+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## master
 
+- Support `stream_id`, `offset`, and `epoch` fields in Action Cable `receive`. ([@palkan][])
+
+- Add `print` option to `receive` action to print the received message. ([@palkan][])
+
 ## 1.0.3 (2023-10-03)
 
 - Add `timeout` option to receive to limit the amount of time we wait for the expected message. ([@palkan][])